Made some changes for solaris build. Modified svkPlotGrid to use a single vtkPolyData and vtkPoints object. Fixed open spectra bug where PTS were selected as the default x range. Fixed flickering bug in OSX and Solaris.

+++ libs/src/svkCoilCombine.cc
@@ -252,8 +252,8 @@ void svkCoilCombine::RequestSumOfSquaresData()
                             magnitudeData[ (freq *2) + 1 ] += 0; 
     
                         }
-
-                        magnitudeData[ freq * 2 ] = pow( magnitudeData[freq * 2], 0.5 );  
+						// Solaris build needs these casts to determine the correct method
+                        magnitudeData[ freq * 2 ] = pow( (double)magnitudeData[freq * 2], (double)0.5 );
                         magnitudeData[ (freq *2) + 1 ] = magnitudeData[ freq * 2 ]; 
 
                     }
